Added email services

+++ src/main/java/org/example/services/UserServiceImpl.java
@@ -2,10 +2,10 @@ package org.example.services;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.example.data.dtos.requests.DeleteUserRequest;
 import org.example.data.models.Roles;
 import org.example.exceptions.UserNotAuthorizedException;
 import org.example.exceptions.UserNotFoundException;
-import org.example.security.AppUtils;
 import org.example.data.dtos.requests.CreateUserRequest;
 import org.example.data.dtos.requests.LoginRequest;
 import org.example.data.dtos.responses.CreateUserResponse;
@@ -17,17 +17,16 @@ import org.example.exceptions.InvalidLoginDetailsException;
 import org.example.exceptions.UsernameAlreadyExistsException;
 import org.example.security.AuthenticatedUser;
 import org.example.security.JwtUtils;
+import org.example.services.cloud.CloudService;
+import org.example.services.email.EmailService;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
+import org.springframework.web.multipart.MultipartFile;
+import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -40,6 +39,8 @@ public class UserServiceImpl implements UserService {
     private final PasswordEncoder passwordEncoder;
     private final AuthenticationManager authenticationManager;
     private final JwtUtils jwtUtils;
+    private final EmailService emailService;
+    private final CloudService cloudService;
 
     @Override
     public CreateUserResponse createUser(CreateUserRequest userRequest){
@@ -54,6 +55,13 @@ public class UserServiceImpl implements UserService {
                 .isEnabled(true)
                 .roles(Set.of(Roles.USER))
                 .build();
+        String htmlContent = "<html>" +
+                "<head></head>" +
+                "<body>" +
+                "<p>Hi there, " + newUser.getFirstName() + "!</p>" +
+                "<p>Welcome to You-RL shortener. We're glad to have you!</p>" +
+                "</body></html>";
+        emailService.sendEmail(newUser.getEmail(), "Welcome to You-RL", htmlContent);
         UserEntity savedUser = userRepository.save(newUser);
         LoginResponse response = this.generateTokens(new HashMap<>(), savedUser.getEmail());
         return CreateUserResponse.builder()
@@ -69,12 +77,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public LoginResponse login(LoginRequest loginRequest) {
         try {
-            log.info("111");
             Authentication authentication = authenticationManager.authenticate(
                     new UsernamePasswordAuthenticationToken(loginRequest.getEmail(),
                             loginRequest.getPassword())
             );
-            log.info("222");
 
             Map<String, Object> claims = authentication.getAuthorities().stream()
                     .collect(Collectors.toMap(authority -> "claim", Function.identity()));
@@ -116,10 +122,41 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public String deleteUserById(Long userId) {
+    public String deleteUser(DeleteUserRequest deleteUserRequest) {
+        Long userId = deleteUserRequest.getUserId();
+        if (!Objects.equals(userId, this.getCurrentUser().getId())) throw new UserNotAuthorizedException();
         UserEntity foundUser = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-        userRepository.delete(foundUser);
+        if (passwordEncoder.matches(deleteUserRequest.getPassword(), foundUser.getPassword())) userRepository.delete(foundUser);
         return "SUCCESSFUL";
+    }
+
+    @Override
+    public String uploadProfileImage(MultipartFile profileImage) {
+        UserEntity foundUser = this.getCurrentUser();
+        String imageUrl = cloudService.upload(profileImage);
+        foundUser.setProfileImage(imageUrl);
+        userRepository.save(foundUser);
+        return imageUrl;
+    }
+
+    @Override
+    public String deleteProfileImage() {
+        UserEntity foundUser = this.getCurrentUser();
+        foundUser.setProfileImage("");
+        userRepository.save(foundUser);
+        return "SUCCESSFUL";
+    }
+
+    @Override
+    public List<FindUserResponse> findAll() {
+        return userRepository.findAll().stream().map((foundUser) -> FindUserResponse.builder()
+                .id(foundUser.getId())
+                .lastName(foundUser.getLastName())
+                .firstName(foundUser.getFirstName())
+                .email(foundUser.getEmail())
+                .username(foundUser.getUsername())
+                .password(foundUser.getPassword())
+                .build()).collect(Collectors.toList());
     }
 
     private LoginResponse generateTokens(Map<String, Object> claims, String email) {
